+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- JPM CEMBI BROAD DIVERSE and PM EMBI GLOBAL DIVERSIFI added to portfolio df
 ### Fixed
 ### Changed
 ### Removed